Remove install function
The idea of the function was to show an example quicktag and have a filter for plugin and theme authors. But the filter was not very useful as it only ran on install - import is better for this.

And the functionality of the example quicktag can be replaced by the placeholder attribute